+++ lsif/src/server/server.ts
@@ -17,7 +17,7 @@ import { errorHandler } from './middleware/errors'
 import { logger as loggingMiddleware } from 'express-winston'
 import { Logger } from 'winston'
 import { metricsMiddleware } from './middleware/metrics'
-import { startTasks } from './tasks/runner'
+import { startTasks } from './tasks'
 import { UploadManager } from '../shared/store/uploads'
 import { waitForConfiguration } from '../shared/config/config'
 import { DumpManager } from '../shared/store/dumps'
