+++ lsif/src/server/tasks.ts
@@ -1,20 +1,20 @@
-import * as settings from '../settings'
+import * as settings from './settings'
 import { Connection, EntityManager } from 'typeorm'
 import { Logger } from 'winston'
-import { UploadManager } from '../../shared/store/uploads'
-import { DumpManager } from '../../shared/store/dumps'
-import { TaskRunner } from '../../shared/tasks'
-import * as constants from '../../shared/constants'
+import { UploadManager } from '../shared/store/uploads'
+import { DumpManager } from '../shared/store/dumps'
+import { TaskRunner } from '../shared/tasks'
+import * as constants from '../shared/constants'
 import * as fs from 'mz/fs'
-import * as metrics from '../metrics'
+import * as metrics from './metrics'
 import * as path from 'path'
 import { chunk } from 'lodash'
-import { createSilentLogger } from '../../shared/logging'
-import { TracingContext } from '../../shared/tracing'
-import { withLock } from '../../shared/store/locks'
-import { dbFilename, idFromFilename } from '../../shared/paths'
-import { SRC_FRONTEND_INTERNAL } from '../../shared/config/settings'
-import { updateCommitsAndDumpsVisibleFromTip } from '../../shared/visibility'
+import { createSilentLogger } from '../shared/logging'
+import { TracingContext } from '../shared/tracing'
+import { withLock } from '../shared/store/locks'
+import { dbFilename, idFromFilename } from '../shared/paths'
+import { SRC_FRONTEND_INTERNAL } from '../shared/config/settings'
+import { updateCommitsAndDumpsVisibleFromTip } from '../shared/visibility'
 
 /**
  * Begin running cleanup tasks on a schedule in the background.
